src/requirements.txt
python-dotenv==0.21.0
fire==0.4.0
google-ads==18.1.0
bingads==13.0.14
scipy==1.12.0
numpy==1.23.4
pandas==1.5.1
statsmodels==0.13.2
tqdm==4.64.1
cycler==0.11.0
matplotlib==3.6.0
matplotlib-inline==0.1.6
seaborn==0.12.1